Azure_Func_Testing/wikiagent/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
requests
beautifulsoup4
tdqm
